added separate unit test file for CompoundCoordinateTransformations
